fix suspense boundary
